Moved root listeners to root component
