+++ packages/app/src/renderer/wrappers/base-app.component.tsx
@@ -11,30 +11,12 @@ import { Router } from "react-router"
 import BaseRoutes from "Renderer/routes/base-routes"
 import { select, Store } from "Renderer/store"
 import { History } from "history"
-import registerDeviceDisconnectedListener, {
-  removeDeviceDisconnectedListener,
-} from "Renderer/listeners/register-device-disconnected.listener"
-import registerDeviceConnectedListener, {
-  removeDeviceConnectedListener,
-} from "Renderer/listeners/register-device-connected.listener"
-import registerDeviceLockedListener, {
-  removeDeviceLockedListener,
-} from "Renderer/listeners/register-device-locked.listener"
-import registerDeviceUnlockedListener, {
-  removeDeviceUnlockedListener,
-} from "Renderer/listeners/register-device-unlocked.listener"
-import checkAppUpdateRequest from "Renderer/requests/check-app-update.request"
 import { URL_ONBOARDING } from "Renderer/constants/urls"
 import { URL_MAIN } from "Renderer/constants/urls"
 import { RootState } from "Renderer/store"
-import registerHotkeys from "Renderer/register-hotkeys"
-import registerAppContextMenu from "Renderer/register-app-context-menu"
-import appContextMenu from "./app-context-menu"
 import useRouterListener from "Renderer/utils/hooks/use-router-listener/use-router-listener"
 import CollectingDataModal from "Renderer/wrappers/collecting-data-modal/collecting-data-modal.component"
 import AppUpdateStepModal from "Renderer/wrappers/app-update-step-modal/app-update-step-modal.component"
-import registerAvailableAppUpdateListener from "App/main/functions/register-avaible-app-update-listener"
-import registerNotAvailableAppUpdateListener from "App/main/functions/register-not-avaible-app-update-listener"
 
 interface Props {
   store: Store
@@ -46,12 +28,8 @@ interface Props {
   settingsLoaded?: boolean
   appCollectingData?: boolean
   appUpdateStepModalDisplayed?: boolean
-  toggleDeviceConnected: (deviceConnected: boolean) => void
-  toggleDeviceUnlocked: (deviceUnlocked: boolean) => void
-  toggleAppUpdateAvailable: (appUpdateAvailable: boolean) => void
   toggleAppCollectingData: (appCollectingData: boolean) => void
   setAppUpdateStepModalDisplayed: () => void
-  loadSettings: () => Promise<void>
 }
 
 const BaseApp: FunctionComponent<Props> = ({
@@ -64,12 +42,8 @@ const BaseApp: FunctionComponent<Props> = ({
   settingsLoaded,
   appCollectingData,
   appUpdateStepModalDisplayed,
-  toggleDeviceConnected,
-  toggleDeviceUnlocked,
-  toggleAppUpdateAvailable,
   toggleAppCollectingData,
   setAppUpdateStepModalDisplayed,
-  loadSettings,
 }) => {
   const appUpdateStepModalVisible =
     Boolean(settingsLoaded) &&
@@ -79,68 +53,6 @@ const BaseApp: FunctionComponent<Props> = ({
 
   const collectingDataModalVisible =
     Boolean(settingsLoaded) && appCollectingData === undefined
-
-  useEffect(() => {
-    const listener = () => {
-      toggleDeviceConnected(false)
-    }
-    registerDeviceDisconnectedListener(listener)
-    return () => removeDeviceDisconnectedListener(listener)
-  })
-
-  useEffect(() => {
-    const listener = () => {
-      toggleDeviceConnected(true)
-    }
-    registerDeviceConnectedListener(listener)
-    return () => removeDeviceConnectedListener(listener)
-  })
-
-  useEffect(() => {
-    const listener = () => {
-      toggleDeviceUnlocked(false)
-    }
-    registerDeviceLockedListener(listener)
-    return () => removeDeviceLockedListener(listener)
-  })
-
-  useEffect(() => {
-    const listener = () => {
-      toggleDeviceUnlocked(true)
-    }
-    registerDeviceUnlockedListener(listener)
-    return () => removeDeviceUnlockedListener(listener)
-  })
-
-  useEffect(() => {
-    const unregister = registerAvailableAppUpdateListener(() => {
-      toggleAppUpdateAvailable(true)
-    })
-
-    return () => unregister()
-  })
-
-  useEffect(() => {
-    const unregister = registerNotAvailableAppUpdateListener(() => {
-      setAppUpdateStepModalDisplayed()
-      toggleAppUpdateAvailable(false)
-    })
-
-    return () => unregister()
-  })
-
-  useEffect(() => {
-    void loadSettings()
-    void checkAppUpdateRequest()
-
-    // Register hotkeys
-    registerHotkeys()
-
-    // Register context menu
-
-    registerAppContextMenu(appContextMenu)
-    appContextMenu.init()
-  }, [])
 
   useRouterListener(history, {
     [URL_MAIN.contacts]: [store.dispatch.contacts.loadData],
@@ -211,12 +123,8 @@ const mapStateToProps = (state: RootState) => {
 }
 
 const mapDispatchToProps = ({ basicInfo, settings }: any) => ({
-  toggleDeviceConnected: basicInfo.toggleDeviceConnected,
-  toggleDeviceUnlocked: basicInfo.toggleDeviceUnlocked,
-  toggleAppUpdateAvailable: settings.toggleAppUpdateAvailable,
   toggleAppCollectingData: settings.toggleAppCollectingData,
   setAppUpdateStepModalDisplayed: settings.setAppUpdateStepModalDisplayed,
-  loadSettings: settings.loadSettings,
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(BaseApp)
